+++ rust/crd/src/authentication.rs
@@ -1,48 +1,59 @@
-use serde::{Deserialize, Serialize};
 use snafu::{ResultExt, Snafu};
 use stackable_operator::{
     client::Client,
-    commons::authentication::AuthenticationClass,
-    kube::runtime::reflector::ObjectRef,
-    schemars::{self, JsonSchema},
+    commons::authentication::{
+        oidc, AuthenticationClass, AuthenticationClassProvider, ClientAuthenticationDetails,
+    },
+    kube::ResourceExt,
 };
 
 #[derive(Snafu, Debug)]
 pub enum Error {
-    #[snafu(display("Failed to retrieve AuthenticationClass {authentication_class}"))]
+    #[snafu(display("Failed to retrieve AuthenticationClass"))]
     AuthenticationClassRetrieval {
         source: stackable_operator::error::Error,
-        authentication_class: ObjectRef<AuthenticationClass>,
+    },
+
+    #[snafu(display("Invalid OIDC configuration"))]
+    InvalidOidcConfiguration {
+        source: stackable_operator::error::Error,
     },
 }
 
 type Result<T, E = Error> = std::result::Result<T, E>;
 
-#[derive(Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TrinoAuthenticationClassRef {
+pub struct ResolvedAuthenticationClassRef {
     /// An [AuthenticationClass](DOCS_BASE_URL_PLACEHOLDER/concepts/authentication) to use.
-    pub authentication_class: String,
+    pub authentication_class: AuthenticationClass,
+    pub oidc: Option<oidc::ClientAuthenticationOptions>,
 }
 
 /// Retrieve all provided AuthenticationClass references.
 pub async fn resolve_authentication_classes(
     client: &Client,
-    authentication_class_refs: &Vec<TrinoAuthenticationClassRef>,
-) -> Result<Vec<AuthenticationClass>> {
+    client_authentication_details: &Vec<ClientAuthenticationDetails>,
+) -> Result<Vec<ResolvedAuthenticationClassRef>> {
     let mut resolved_auth_classes = vec![];
 
-    for auth_class in authentication_class_refs {
-        let resolved_auth_class =
-            AuthenticationClass::resolve(client, &auth_class.authentication_class)
-                .await
-                .context(AuthenticationClassRetrievalSnafu {
-                    authentication_class: ObjectRef::<AuthenticationClass>::new(
-                        &auth_class.authentication_class,
-                    ),
-                })?;
+    for client_authentication_detail in client_authentication_details {
+        let resolved_auth_class = client_authentication_detail
+            .resolve_class(client)
+            .await
+            .context(AuthenticationClassRetrievalSnafu)?;
+        let auth_class_name = resolved_auth_class.name_any();
 
-        resolved_auth_classes.push(resolved_auth_class);
+        resolved_auth_classes.push(ResolvedAuthenticationClassRef {
+            oidc: match &resolved_auth_class.spec.provider {
+                AuthenticationClassProvider::Oidc(_) => Some(
+                    client_authentication_detail
+                        .oidc_or_error(&auth_class_name)
+                        .context(InvalidOidcConfigurationSnafu)?
+                        .clone(),
+                ),
+                _ => None,
+            },
+            authentication_class: resolved_auth_class,
+        });
     }
 
     Ok(resolved_auth_classes)